foodgram/api/mixins.py
from rest_framework import mixins, viewsets


class CreateDestroyViewSet(
    mixins.CreateModelMixin,
    mixins.DestroyModelMixin,
    viewsets.GenericViewSet
):
     """Кастомный миксин для ViewSet c Create, Destroy методами."""
