requirements.txt
Django==2.2.19
djangorestframework
djoser
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
